requirements.txt
pytorch-lightning
torch_gpu